% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_knockoffs.R
\name{add_knockoffs}
\alias{add_knockoffs}
\title{Add knockoffs to the original predictor matrix}
\usage{
add_knockoffs(
  X,
  L_val,
  cor.structure = FALSE,
  empirical = FALSE,
  eps = .Machine$double.eps
)
}
\arguments{
\item{X}{Real valued Predictor matrix.}

\item{L_val}{Number of knockoffs.}

\item{cor.structure}{TRUE/FALSE.}

\item{empirical}{TRUE/FALSE.}

\item{eps}{Numerical zero.}
}
\value{
Enlarged predictor matrix, i.e., original predictor matrix and knockoffs.
}
\description{
Sample L_val knockoff vectors from the univariate standard normal distribution and append them to the predictor matrix X.
}
\examples{
n = 50
p = 100
add_knockoffs(X = matrix(rnorm(n * p), nrow = n, ncol = p), L_val = p)
}
