% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_knockoffs_GVS.R
\name{add_knockoffs_GVS}
\alias{add_knockoffs_GVS}
\title{Add GVS-knockoffs to the original predictor matrix}
\usage{
add_knockoffs_GVS(X, L_val, corr_max = 0.5)
}
\arguments{
\item{X}{Real valued predictor matrix.}

\item{L_val}{Number of knockoffs.}

\item{corr_max}{Maximum allowed correlation between any two predictors from different clusters.}
}
\value{
Enlarged predictor matrix for the T-Knock+GVS filter, i.e., original predictor matrix and knockoffs.
}
\description{
Sample L_val knockoff vectors for the T-Knock+GVS filter from the univariate standard normal distribution and append them to the predictor matrix X.
}
\examples{
n = 50
p = 100
add_knockoffs_GVS(X = matrix(rnorm(n * p), nrow = n, ncol = p), L_val = p)
}
