% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{agglomerative_cluster}
\alias{agglomerative_cluster}
\title{Agglomerative Clustering}
\usage{
agglomerative_cluster(
  data,
  method = LinkageMethod$Ward,
  metric = DistanceMetric$Euclidean,
  use_mmap = FALSE
)
}
\arguments{
\item{data}{A numeric matrix.}

\item{method}{A value from \code{LinkageMethod} (e.g. \code{LinkageMethod$Ward}).}

\item{metric}{A value from \code{DistanceMetric} (e.g. \code{DistanceMetric$Euclidean}).
Default is Euclidean.}

\item{use_mmap}{Whether to use memory mapping files internally. Default is FALSE.}
}
\value{
A numeric matrix representing the computed hierarchical linkage.
}
\description{
Performs hierarchical agglomerative clustering on a numeric matrix.
}
