% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{convert_to_memory_mapped}
\alias{convert_to_memory_mapped}
\title{Convert R Matrix to System Memory Mapped Matrix}
\usage{
convert_to_memory_mapped(mat, filename)
}
\arguments{
\item{mat}{A \code{numeric} \code{matrix}.}

\item{filename}{String destination file path.}
}
\value{
NULL invisibly mapped.
}
\description{
Maps a given R numeric matrix into a binary formatted memory-mapped file on disk
efficiently.
}
