% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDP.R
\name{FDP}
\alias{FDP}
\title{False discovery proportion (FDP)}
\usage{
FDP(beta_hat, beta, eps = .Machine$double.eps)
}
\arguments{
\item{beta_hat}{Estimated regression coefficient vector.}

\item{beta}{True regression coefficient vector.}

\item{eps}{Numerical zero.}
}
\value{
False discovery proportion (FDP).
}
\description{
Computes the FDP based on the estimated and the true regression coefficient vectors.
}
\examples{
data("Gauss_data")
X <- Gauss_data$X
y <- c(Gauss_data$y)
beta <- Gauss_data$beta

set.seed(1234)
res <- tknock(X, y)
beta_hat <- res$selected_var

FDP(beta_hat = beta_hat, beta = beta)
}
