% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fdp_hat.R
\name{fdp_hat}
\alias{fdp_hat}
\title{Computes the conservative FDP estimate of the T-Knock filter}
\usage{
fdp_hat(V, Phi, Phi_prime, T_stop, L_val, eps = .Machine$double.eps)
}
\arguments{
\item{V}{Voting level grid.}

\item{Phi}{Vector of relative occurrences.}

\item{Phi_prime}{Vector of deflated relative occurrences.}

\item{T_stop}{Number of included knockoffs after which the random experiments (i.e., forward selection processes) are stopped.}

\item{L_val}{Number of knockoffs.}

\item{eps}{Numerical zero.}
}
\value{
Vector of conservative FDP estimates for each value of the voting level grid.
}
\description{
Computes the conservative FDP estimate of the T-Knock filter
}
