% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_ptr}
\alias{get_ptr}
\title{Get the External Pointer of a MemoryMappedMatrix}
\usage{
get_ptr(x, ...)
}
\arguments{
\item{x}{A \code{MemoryMappedMatrix} / \code{mmap_matrix} object.}

\item{...}{unused arguments}
}
\value{
The underlying external pointer (the object itself).
}
\description{
Get the External Pointer of a MemoryMappedMatrix
}
