% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ground_truth.R
\docType{data}
\name{ground_truth}
\alias{ground_truth}
\title{Ground truth for simulated GWAS}
\format{
A dataframe containing six columns:
\describe{
\item{allele}{Disease allele (zero or one).}
\item{idx}{Index of disease SNP.}
\item{position}{Position of disease SNP.}
\item{risk_hete}{Heterozygote risk.}
\item{risk_homo}{Homozygote risk.}
\item{rs}{rs-ID-number.}
}
}
\usage{
ground_truth
}
\description{
A table containing the disease SNPs (disease allele, index, position, heterozygote risks,
homozygote risks, and rs-ID-number).
}
\keyword{datasets}
