% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Phi_prime_fun.R
\name{Phi_prime_fun}
\alias{Phi_prime_fun}
\title{Computes the Deflated Relative Occurrences}
\usage{
Phi_prime_fun(p, T_stop, L_val, phi_T_mat, Phi, eps = .Machine$double.eps)
}
\arguments{
\item{p}{Number of candidate variables.}

\item{T_stop}{Number of included knockoffs after which the random experiments (i.e., forward selection processes) are stopped.}

\item{L_val}{Number of knockoffs.}

\item{phi_T_mat}{Matrix of relative occurrences for all variables (i.e., j = 1,..., p) and for T = 1, ..., T_stop.}

\item{Phi}{Vector of relative occurrences for all variables (i.e., j = 1,..., p) at T = T_stop.}

\item{eps}{Numerical zero.}
}
\value{
Matrix of deflated relative occurrences for all variables (i.e., j = 1,..., p) and for T = 1, ..., T_stop.
}
\description{
Computes the matrix of deflated relative occurrences for all variables (i.e., j = 1,..., p) and for T = 1, ..., T_stop.
}
\examples{
data("Gauss_data")
X = Gauss_data$X
y = c(Gauss_data$y)
rand_exp = random_experiments(X, y)
Phi_prime_fun(p = ncol(X),
              T_stop = rand_exp$T_stop,
              L_val = rand_exp$L_val,
              phi_T_mat = rand_exp$phi_T_mat,
              Phi = rand_exp$Phi,
              eps = rand_exp$eps)
}
