% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.mmap_matrix}
\alias{print.mmap_matrix}
\title{Print mmap_matrix}
\usage{
\method{print}{mmap_matrix}(x, ...)
}
\arguments{
\item{x}{mmap_matrix object}

\item{...}{arguments passed to other methods}
}
\value{
invisibly returns x
}
\description{
Print mmap_matrix
}
