% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_var_fun.R
\name{select_var_fun}
\alias{select_var_fun}
\title{Compute set of selected variables}
\usage{
select_var_fun(p, tFDR, T_stop, FDP_hat_mat, Phi_mat, V)
}
\arguments{
\item{p}{Number of candidate variables.}

\item{tFDR}{Target FDR level (between 0 and 1, i.e., 0\% and 100\%).}

\item{T_stop}{Number of included dummies after which the random experiments (i.e., forward selection processes) are stopped.}

\item{FDP_hat_mat}{Matrix whose rows are the vectors of conservative FDP estimates for each value of the voting level grid.}

\item{Phi_mat}{Matrix of relative occurrences as determined by the T-Rex calibration algorithm.}

\item{V}{Voting level grid.}
}
\value{
Estimated support vector.
}
\description{
Computes the set of selected variables and returns the estimated support vector for the T-Rex selector.
}
