% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_var_fun_DA_BT.R
\name{select_var_fun_DA_BT}
\alias{select_var_fun_DA_BT}
\title{Compute set of selected variables for the T-Rex+DA+BT selector T-Rex+DA+BT selector (\doi{10.48550/arXiv.2401.15796})}
\usage{
select_var_fun_DA_BT(
  p,
  tFDR,
  T_stop,
  FDP_hat_array_BT,
  Phi_array_BT,
  V,
  rho_grid
)
}
\arguments{
\item{p}{Number of candidate variables.}

\item{tFDR}{Target FDR level (between 0 and 1, i.e., 0\% and 100\%).}

\item{T_stop}{Number of included dummies after which the random experiments (i.e., forward selection processes) are stopped.}

\item{FDP_hat_array_BT}{Array containing the conservative FDP estimates for all variables (dimension 1),
values of the voting level grid (dimension 2), and values of the dendrogram grid (dimension 3).}

\item{Phi_array_BT}{Array of relative occurrences as determined by the T-Rex calibration algorithm.}

\item{V}{Voting level grid.}

\item{rho_grid}{Dendrogram grid.}
}
\value{
List containing the estimated support vector, etc.
}
\description{
Computes the set of selected variables and returns the estimated support vector for the T-Rex+DA+BT selector (\doi{10.48550/arXiv.2401.15796}).
}
