% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_cases.R
\docType{data}
\name{sim_data_cases}
\alias{sim_data_cases}
\title{Genotypes for cases generated with HAPGEN2 using haplotypes from HapMap 3}
\format{
An object of the class "SnpMatrix".
}
\usage{
sim_data_cases
}
\description{
A matrix containing 1000 SNPs of 500 cases.
}
\keyword{datasets}
