% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_data_controls.R
\docType{data}
\name{sim_data_controls}
\alias{sim_data_controls}
\title{Genotypes for controls generated with HAPGEN2 using haplotypes from HapMap 3}
\format{
An object of the class "SnpMatrix".
}
\usage{
sim_data_controls
}
\description{
A matrix containing 1000 SNPs of 200 controls.
}
\keyword{datasets}
