% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData_cases.R
\docType{data}
\name{simData_cases}
\alias{simData_cases}
\title{Genotypes for cases}
\format{
An object of the class "SnpMatrix".
}
\usage{
simData_cases
}
\description{
A matrix containing the SNPs of the cases.
}
\keyword{datasets}
