% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsolvers.R
\name{TAFSSolver}
\alias{TAFSSolver}
\title{T-AFS Solver}
\description{
Terminating Adaptive Forward Stepwise (T-AFS) solver.
The shrinkage parameter \code{rho} interpolates between Forward Stepwise
(\code{rho = 1}) and LAR/LASSO (\code{rho -> 0}).
Inherits all common methods from \code{\link{TSolverBase}}.
}
\examples{
set.seed(1)
n <- 20; p <- 5
X <- matrix(rnorm(n * p), n, p)
D <- matrix(rnorm(n * p), n, p)
y <- X[, 1] * 2 + rnorm(n)
solver <- TAFSSolver$new(X, D, y, rho = 0.5)
solver$execute_step(T_stop = 3)
solver$get_beta()

}
\section{Super class}{
\code{\link[TRexSelector:TSolverBase]{TRexSelector::TSolverBase}} -> \code{TAFSSolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TAFSSolver-new}{\code{TAFSSolver$new()}}
\item \href{#method-TAFSSolver-clone}{\code{TAFSSolver$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="execute_step"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-execute_step'><code>TRexSelector::TSolverBase$execute_step()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_actions"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_actions'><code>TRexSelector::TSolverBase$get_actions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_actives"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_actives'><code>TRexSelector::TSolverBase$get_actives()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_beta"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_beta'><code>TRexSelector::TSolverBase$get_beta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_cp"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_cp'><code>TRexSelector::TSolverBase$get_cp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_dof"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_dof'><code>TRexSelector::TSolverBase$get_dof()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_inactives"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_inactives'><code>TRexSelector::TSolverBase$get_inactives()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_intercept"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_intercept'><code>TRexSelector::TSolverBase$get_intercept()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_num_steps"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_num_steps'><code>TRexSelector::TSolverBase$get_num_steps()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_r2"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_r2'><code>TRexSelector::TSolverBase$get_r2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_rss"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_rss'><code>TRexSelector::TSolverBase$get_rss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="load"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-load'><code>TRexSelector::TSolverBase$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="save"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-save'><code>TRexSelector::TSolverBase$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TAFSSolver-new"></a>}}
\if{latex}{\out{\hypertarget{method-TAFSSolver-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TAFSSolver}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TAFSSolver$new(
  X,
  D,
  y,
  rho = 1,
  normalize = TRUE,
  intercept = TRUE,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Numeric matrix of predictors, or \code{mmap_matrix}.}

\item{\code{D}}{Numeric matrix of dummies.}

\item{\code{y}}{Numeric response vector.}

\item{\code{rho}}{Numeric shrinkage parameter in (0, 1]; default \code{1.0}.}

\item{\code{normalize}}{Logical; column-normalize X and D (default \code{TRUE}).}

\item{\code{intercept}}{Logical; fit an intercept (default \code{TRUE}).}

\item{\code{verbose}}{Logical; print progress (default \code{FALSE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TAFSSolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TAFSSolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TAFSSolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
