% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TPP.R
\name{TPP}
\alias{TPP}
\title{True positive proportion (TPP)}
\usage{
TPP(beta_hat, beta, eps = .Machine$double.eps)
}
\arguments{
\item{beta_hat}{Estimated support vector.}

\item{beta}{True support vector.}

\item{eps}{Numerical zero.}
}
\value{
TPP.
}
\description{
True positive proportion (TPP)
}
