% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/control_parameters.R
\name{trex_screen_control}
\alias{trex_screen_control}
\title{TRex Screen Control Parameters}
\usage{
trex_screen_control(
  use_bootstrap_CI = FALSE,
  R_boot = 1000,
  ci_grid_step = 0.001,
  trex_method = "TREX"
)
}
\arguments{
\item{use_bootstrap_CI}{Whether to use bootstrap confidence intervals for screening
(default: FALSE).}

\item{R_boot}{Number of bootstrap iterations (default: 1000).}

\item{ci_grid_step}{Grid step size for confidence-interval calculations (default: 0.001).}

\item{trex_method}{Screening method: "TREX", "TREX_DA_AR1", "TREX_DA_EQUI",
"TREX_DA_BLOCK_EQUI" (default: "TREX").}
}
\value{
A named list for use in TRexScreeningSelector or TRexBiobankScreeningSelector.
}
\description{
Build a control list for the screening step in TRexScreeningSelector and
TRexBiobankScreeningSelector.
}
\examples{
trex_screen_control()
trex_screen_control(use_bootstrap_CI = TRUE, R_boot = 500)

}
