% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trex_screening_selector.R
\name{TRexScreeningSelector}
\alias{TRexScreeningSelector}
\title{Screen T-Rex Selector}
\description{
Fast variable screening algorithm based on TRex architecture.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TRexScreeningSelector-new}{\code{TRexScreeningSelector$new()}}
\item \href{#method-TRexScreeningSelector-select}{\code{TRexScreeningSelector$select()}}
\item \href{#method-TRexScreeningSelector-get_selected_indices}{\code{TRexScreeningSelector$get_selected_indices()}}
\item \href{#method-TRexScreeningSelector-get_results}{\code{TRexScreeningSelector$get_results()}}
\item \href{#method-TRexScreeningSelector-get_matrices}{\code{TRexScreeningSelector$get_matrices()}}
\item \href{#method-TRexScreeningSelector-clone}{\code{TRexScreeningSelector$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRexScreeningSelector-new"></a>}}
\if{latex}{\out{\hypertarget{method-TRexScreeningSelector-new}{}}}
\subsection{Method \code{new()}}{
Create a new TRexScreeningSelector object.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRexScreeningSelector$new(
  X,
  y,
  seed = -1,
  verbose = TRUE,
  use_bootstrap_CI = FALSE,
  R_boot = 1000,
  ci_grid_step = 0.001,
  trex_method = "TREX",
  method = "LARS",
  use_omp = TRUE,
  use_memory_mapping = FALSE,
  max_outer_threads = 1,
  max_inner_threads = 1
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Feature matrix or MemoryMappedMatrix.}

\item{\code{y}}{Response vector.}

\item{\code{seed}}{Random seed (default: -1).}

\item{\code{verbose}}{Whether to print progress (default: TRUE).}

\item{\code{use_bootstrap_CI}}{Whether to use bootstrap CI screening.}

\item{\code{R_boot}}{Number of bootstrap iterations.}

\item{\code{ci_grid_step}}{the grid step sizes for interval calculations}

\item{\code{trex_method}}{Screening method: "TREX", "TREX_DA_AR1", "TREX_DA_EQUI", "TREX_DA_BLOCK_EQUI"}

\item{\code{method}}{The solver method, e.g., "LARS", "TACGP", "TENET" (default: "LARS").}

\item{\code{use_omp}}{Whether to use OpenMP parallelization (default: TRUE).}

\item{\code{use_memory_mapping}}{Use out-of-core memory mapping (default: FALSE).}

\item{\code{max_outer_threads}}{Thread count for K repetitions (default: 1).}

\item{\code{max_inner_threads}}{Thread count for inner solvers (default: 1).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRexScreeningSelector-select"></a>}}
\if{latex}{\out{\hypertarget{method-TRexScreeningSelector-select}{}}}
\subsection{Method \code{select()}}{
Run the Screening algorithm.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRexScreeningSelector$select()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRexScreeningSelector-get_selected_indices"></a>}}
\if{latex}{\out{\hypertarget{method-TRexScreeningSelector-get_selected_indices}{}}}
\subsection{Method \code{get_selected_indices()}}{
Get the integer indices (1-based) of the selected variables.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRexScreeningSelector$get_selected_indices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric vector of selected indices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRexScreeningSelector-get_results"></a>}}
\if{latex}{\out{\hypertarget{method-TRexScreeningSelector-get_results}{}}}
\subsection{Method \code{get_results()}}{
Get generic solver results.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRexScreeningSelector$get_results()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing results.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRexScreeningSelector-get_matrices"></a>}}
\if{latex}{\out{\hypertarget{method-TRexScreeningSelector-get_matrices}{}}}
\subsection{Method \code{get_matrices()}}{
Get solver matrices.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRexScreeningSelector$get_matrices()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
A list containing matrices.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TRexScreeningSelector-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TRexScreeningSelector-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TRexScreeningSelector$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
