% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsolvers.R
\name{TStagewiseSolver}
\alias{TStagewiseSolver}
\title{T-Stagewise Solver}
\description{
Terminating Stagewise Regression solver.
Inherits all common methods from \code{\link{TSolverBase}}.
}
\section{Super class}{
\code{\link[TRexSelector:TSolverBase]{TRexSelector::TSolverBase}} -> \code{TStagewiseSolver}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TStagewiseSolver-new}{\code{TStagewiseSolver$new()}}
\item \href{#method-TStagewiseSolver-get_num_removals}{\code{TStagewiseSolver$get_num_removals()}}
\item \href{#method-TStagewiseSolver-get_cycling_ratio}{\code{TStagewiseSolver$get_cycling_ratio()}}
\item \href{#method-TStagewiseSolver-get_lambda}{\code{TStagewiseSolver$get_lambda()}}
\item \href{#method-TStagewiseSolver-clone}{\code{TStagewiseSolver$clone()}}
}
}
\if{html}{\out{
<details><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="execute_step"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-execute_step'><code>TRexSelector::TSolverBase$execute_step()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_actions"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_actions'><code>TRexSelector::TSolverBase$get_actions()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_actives"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_actives'><code>TRexSelector::TSolverBase$get_actives()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_beta"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_beta'><code>TRexSelector::TSolverBase$get_beta()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_cp"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_cp'><code>TRexSelector::TSolverBase$get_cp()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_dof"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_dof'><code>TRexSelector::TSolverBase$get_dof()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_inactives"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_inactives'><code>TRexSelector::TSolverBase$get_inactives()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_intercept"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_intercept'><code>TRexSelector::TSolverBase$get_intercept()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_num_steps"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_num_steps'><code>TRexSelector::TSolverBase$get_num_steps()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_r2"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_r2'><code>TRexSelector::TSolverBase$get_r2()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="get_rss"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-get_rss'><code>TRexSelector::TSolverBase$get_rss()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="load"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-load'><code>TRexSelector::TSolverBase$load()</code></a></span></li>
<li><span class="pkg-link" data-pkg="TRexSelector" data-topic="TSolverBase" data-id="save"><a href='../../TRexSelector/html/TSolverBase.html#method-TSolverBase-save'><code>TRexSelector::TSolverBase$save()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TStagewiseSolver-new"></a>}}
\if{latex}{\out{\hypertarget{method-TStagewiseSolver-new}{}}}
\subsection{Method \code{new()}}{
Create a new \code{TStagewiseSolver}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TStagewiseSolver$new(
  X,
  D,
  y,
  normalize = TRUE,
  intercept = TRUE,
  verbose = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{X}}{Numeric matrix of predictors, or \code{mmap_matrix}.}

\item{\code{D}}{Numeric matrix of dummies.}

\item{\code{y}}{Numeric response vector.}

\item{\code{normalize}}{Logical; column-normalize X and D (default \code{TRUE}).}

\item{\code{intercept}}{Logical; fit an intercept (default \code{TRUE}).}

\item{\code{verbose}}{Logical; print progress (default \code{FALSE}).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TStagewiseSolver-get_num_removals"></a>}}
\if{latex}{\out{\hypertarget{method-TStagewiseSolver-get_num_removals}{}}}
\subsection{Method \code{get_num_removals()}}{
Get the number of coefficient removals.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TStagewiseSolver$get_num_removals()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Integer count of removals.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TStagewiseSolver-get_cycling_ratio"></a>}}
\if{latex}{\out{\hypertarget{method-TStagewiseSolver-get_cycling_ratio}{}}}
\subsection{Method \code{get_cycling_ratio()}}{
Get the cycling ratio (removals / total steps).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TStagewiseSolver$get_cycling_ratio()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Scalar numeric cycling ratio.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TStagewiseSolver-get_lambda"></a>}}
\if{latex}{\out{\hypertarget{method-TStagewiseSolver-get_lambda}{}}}
\subsection{Method \code{get_lambda()}}{
Get the regularization (lambda) path.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TStagewiseSolver$get_lambda()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
Numeric vector of lambda values, one per step.
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TStagewiseSolver-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TStagewiseSolver-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TStagewiseSolver$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
