% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummies.R
\name{add_dummies}
\alias{add_dummies}
\title{Add dummies to the original predictor matrix}
\usage{
add_dummies(
  X,
  num_dummies,
  cor.structure = FALSE,
  empirical = FALSE,
  eps = .Machine$double.eps
)
}
\arguments{
\item{X}{Real valued predictor matrix.}

\item{num_dummies}{Number of dummies that are appended to the predictor matrix.}

\item{cor.structure}{TRUE/FALSE.}

\item{empirical}{TRUE/FALSE.}

\item{eps}{Numerical zero.}
}
\value{
Enlarged predictor matrix, i.e., original predictor matrix and dummies
}
\description{
Sample num_dummies dummy vectors from the univariate standard normal distribution and append them to the predictor matrix X.
}
\examples{
n <- 50
p <- 100
add_dummies(X = matrix(rnorm(n * p), nrow = n, ncol = p), num_dummies = p)
}
