% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dummies_GVS.R
\name{add_dummies_GVS}
\alias{add_dummies_GVS}
\title{Add GVS-dummies to the original predictor matrix}
\usage{
add_dummies_GVS(X, num_dummies, corr_max = 0.5)
}
\arguments{
\item{X}{Real valued predictor matrix.}

\item{num_dummies}{Number of dummies.}

\item{corr_max}{Maximum allowed correlation between any two predictors from different clusters.}
}
\value{
Enlarged predictor matrix for the T-Knock+GVS filter, i.e., original predictor matrix and dummies
}
\description{
Sample num_dummies dummy vectors for the T-Knock+GVS filter from the univariate standard normal distribution and append them to the predictor matrix X.
}
\examples{
set.seed(123)
n <- 50
p <- 100
add_dummies_GVS(X = matrix(rnorm(n * p), nrow = n, ncol = p), num_dummies = p)
}
