% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/FDP.R
\name{FDP}
\alias{FDP}
\title{False discovery proportion (FDP)}
\usage{
FDP(beta_hat, beta, eps = .Machine$double.eps)
}
\arguments{
\item{beta_hat}{Estimated support vector.}

\item{beta}{True support vector.}

\item{eps}{Numerical zero.}
}
\value{
FDP
}
\description{
False discovery proportion (FDP)
}
