% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_experiments.R
\name{random_experiments}
\alias{random_experiments}
\title{Runs K random experiments and computes the matrix of relative occurrences}
\usage{
random_experiments(
  X,
  y,
  K = 20,
  T_stop = 1,
  L_val = ncol(X),
  method = "tknock",
  type = "lar",
  corr_max = 0.5,
  lambda_2_lars = NULL,
  earlyStop = TRUE,
  lars_state_list,
  verbose = TRUE,
  intercept = FALSE,
  normalize = TRUE,
  cor.structure = FALSE,
  empirical = TRUE,
  parallel_process = FALSE,
  parallel_max_cores = min(K, max(1, parallel::detectCores(logical = FALSE))),
  seed = NULL,
  eps = .Machine$double.eps
)
}
\arguments{
\item{X}{Real valued Predictor matrix.}

\item{y}{Response vector.}

\item{K}{Number of random experiments.}

\item{T_stop}{Number of included dummies after which the random experiments (i.e., forward selection processes) are stopped.}

\item{L_val}{Number of dummies}

\item{method}{'tknock' for T-Knock filter and 'tknock+GVS' for T-Knock+GVS filter.}

\item{type}{'lar' for 'LARS' and 'lasso' for Lasso.}

\item{corr_max}{Maximum allowed correlation between any two predictors from different clusters.}

\item{lambda_2_lars}{lambda_2-value for LARS-based Elastic Net.}

\item{earlyStop}{If TRUE, then the forward selection process is stopped after T_stop dummies have been included. Otherwise the entire solution path is computed.}

\item{lars_state_list}{List of variables associated with previous stopping points of the K random experiments (necessary to restart forward selection selection process exactly where it was previously terminated).}

\item{verbose}{If TRUE progress in computations is shown.}

\item{intercept}{If TRUE an intercept is included.}

\item{normalize}{If TRUE the predictors are standardized and the response is centered.}

\item{cor.structure}{TRUE/FALSE.}

\item{empirical}{TRUE/FALSE.}

\item{parallel_process}{If TRUE random experiments are executed in parallel.}

\item{parallel_max_cores}{Maximum number of cores to be used for parallel processing (default: number of physical cores).}

\item{seed}{Seed for random number generator (only if parallel_process = TRUE).}

\item{eps}{Numerical zero.}
}
\value{
Results of the K random experiments.
}
\description{
Runs K random experiments and computes the matrix of relative occurrences
}
\examples{
set.seed(123)
data('Gauss_data')
X = Gauss_data$X
y = c(Gauss_data$y)
res = random_experiments(X, y)
res$phi_T_mat
}
