% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simData_controls.R
\docType{data}
\name{sim_data_controls}
\alias{sim_data_controls}
\title{Genotypes for controls.}
\format{
An object of the class "SnpMatrix".
}
\usage{
sim_data_controls
}
\description{
A matrix containing the SNPs of the controls.
}
\keyword{datasets}
