% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tknock.R
\name{tknock}
\alias{tknock}
\title{Runs the T-Knock filter with the extended calibration algorithm}
\usage{
tknock(
  X,
  y,
  tFDR = 0.2,
  K = 20,
  max_LL = 10,
  max_T_stop = TRUE,
  method = "tknock",
  type = "lar",
  corr_max = 0.5,
  lambda_2_lars = NULL,
  parallel_process = FALSE,
  parallel_max_cores = min(K, max(1, parallel::detectCores(logical = FALSE))),
  seed = NULL,
  eps = .Machine$double.eps,
  verbose = TRUE
)
}
\arguments{
\item{X}{Real valued predictor matrix.}

\item{y}{Response vector.}

\item{tFDR}{Target FDR level (between 0 and 1, i.e., 0\% and 100\%).}

\item{K}{Number of random experiments.}

\item{max_LL}{Integer factor determining the maximum number of knockoffs as a multiple of p (i.e., L_val = max_LL * p).}

\item{max_T_stop}{If TRUE the maximum number of knockoffs that can be included before stopping is set to ceiling(n / 2).}

\item{method}{'tknock' for T-Knock filter and 'tknock+GVS' for T-Knock+GVS filter.}

\item{type}{'lar' for 'LARS' and 'lasso' for Lasso.}

\item{corr_max}{Maximum allowed correlation between any two predictors from different clusters.}

\item{lambda_2_lars}{lambda_2-value for LARS-based Elastic Net.}

\item{parallel_process}{If TRUE random experiments are executed in parallel.}

\item{parallel_max_cores}{Maximum number of cores to be used for parallel processing (default: number of physical cores).}

\item{seed}{Seed for random number generator (only if parallel_process = TRUE).}

\item{eps}{Numerical zero.}

\item{verbose}{If TRUE progress in computations is shown.}
}
\value{
A list containing the support vector, and the values of T and L as determined by the extended calibration algorithm.
}
\description{
The T-Knock filter performs high-dimensional variable selection for a user-defined target false discovery rate (target FDR).
}
\examples{
data("Gauss_data")
X = Gauss_data$X
y = c(Gauss_data$y)
set.seed(1234)
res = tknock(X, y)
res$beta.selected
}
